moved Metrics from outputs to dataconnector
